Made a self-extracting configuration procedure.
If the configuration files are missing they are automatically recreated from within the JAR file.

Moved all the configurations in to the conf subfolder. Made it so it should work both from within Eclipse and the exported JAR file.

It is important to export the JAR as "extract all libraries..." instead of "package libraries..." to avoid the eclipse Jar-In-Jar loader from messing up this whole thing!