Added first entity configuration.
For now, only minecarts configured. 

An entity is defined by it's name, it has a handler (to interpret the data) a model and materials (like blocks), but no data/mask (handler deals with that) or occlusion.

Reuse of org.jmc.models was attempted. Since the addModel method in blocks uses ints, a new interface was added with method addEntity. For now two models are usable for entities: Minecart and Mesh.

One thing not working correctly is the rotation. For some reason, Minecraft writes bogus values for rotation (at least for me).

In the future, entity models should be added to the final OBJ in such a way that they can be split into individual objects if necessary in order to be easily manipulated in 3D programs.